Separated into basic folders
